MCP Server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu==1.11.0",
    "fastapi>=0.121.2",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-openai==0.3.17",
    "mcp[cli]>=1.21.0",
    "mysql-connector-python>=9.5.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
